Fix various compiler warnings on 64-bit and 32-bit platforms; Also fix build errors on 32-bit platforms
64-bit platform warnings
------------------------
* The below [-Wpedantic] warning was fixed by treating the 128-bit hash as two 64-bit hashes and
  doing 64-bit arithmetic (modulo and divide operations) on those two 64-bit numbers separately.

  ```
  [ 27%] Building C object src/CMakeFiles/libocto.dir/ydb_hash_to_string.c.o
  src/ydb_hash_to_string.c: In function 'ydb_hash_to_string':
  src/ydb_hash_to_string.c:28:11: warning: ISO C does not support '__int128' types [-Wpedantic]
    unsigned __int128 *long_hash;
          ^~~~~~~~
  src/ydb_hash_to_string.c:29:24: warning: ISO C does not support '__int128' types [-Wpedantic]
    long_hash = (unsigned __int128*)hash;
                       ^~~~~~~~
  ```

* The below [-Wunused-function] warnings are fixed by adding "%option noyy_top_state" etc. in lexer.l.

  ```
  [ 56%] Building C object src/CMakeFiles/libocto.dir/lexer.c.o
  build/src/lexer.c:3190:16: warning: 'yy_top_state' defined but not used [-Wunused-function]
       static int yy_top_state  (yyscan_t yyscanner)
               ^~~~~~~~~~~~
  build/src/lexer.c:3181:17: warning: 'yy_pop_state' defined but not used [-Wunused-function]
       static void yy_pop_state  (yyscan_t yyscanner)
                ^~~~~~~~~~~~
  build/src/lexer.c:3155:17: warning: 'yy_push_state' defined but not used [-Wunused-function]
       static void yy_push_state (int  _new_state , yyscan_t yyscanner)
                ^~~~~~~~~~~~~
  build/src/lexer.c:2726:16: warning: 'input' defined but not used [-Wunused-function]
       static int input  (yyscan_t yyscanner)
               ^~~~~
  build/src/lexer.c:2682:17: warning: 'yyunput' defined but not used [-Wunused-function]
       static void yyunput (int c, char * yy_bp , yyscan_t yyscanner)
                ^~~~~~~
  ```

* There is still one type of compiler warning (shown below) that is not fixed on 64-bit platforms.
  That will be attempted at a later point in time.

  ```
  [ 25%] [BISON][OctoParser] Building parser with bison 3.0.4
  build/parser.y: warning: 52 shift/reduce conflicts [-Wconflicts-sr]
  ```

  The reference file that has list of expected warnings is updated to reflect only this warning.

32-bit platform warnings/errors
--------------------------------
* The code in src/ydb_hash_to_string.c had the following build error.  This is automatically fixed
  as part of reworking the logic to treat a 128-bit number as two 64-bit numbers.

  ```
  [ 42%] Building C object src/CMakeFiles/libocto.dir/ydb_hash_to_string.c.o
  src/ydb_hash_to_string.c: In function 'ydb_hash_to_string':
  src/ydb_hash_to_string.c:32:29: error: 'hash' is a pointer; did you mean to use '->'?
    long_hash = (int64_t*)&hash.one;
                            ^
                            ->
  ```

* The below [-Wformat=] warnings are fixed by typecasting value to be "long long int" which translates
  to an 8-byte quantity on both 64-bit and 32-bit platforms and using "%lld" instead of "%ld" in the
  snprintf.

  ```
  [  1%] Building C object src/CMakeFiles/librocto.dir/rocto/copy_bind_parameter.c.o
  src/rocto/copy_bind_parameter.c: In function 'copy_binary_parameter':
  src/rocto/copy_bind_parameter.c:64:51: warning: format '%ld' expects argument of type 'long int', but argument 4 has type 'int64_t {aka long long int}' [-Wformat=]
      copied = snprintf(str_value, MAX_STR_CONST, "%ld", value);
                                                  ^
  src/rocto/copy_bind_parameter.c:68:51: warning: format '%ld' expects argument of type 'long int', but argument 4 has type 'int64_t {aka long long int}' [-Wformat=]
      copied = snprintf(str_value, MAX_STR_CONST, "%ld", value);
                                                  ^
  src/rocto/copy_bind_parameter.c:72:51: warning: format '%ld' expects argument of type 'long int', but argument 4 has type 'int64_t {aka long long int}' [-Wformat=]
      copied = snprintf(str_value, MAX_STR_CONST, "%ld", value);
                                                  ^
  src/rocto/copy_bind_parameter.c:77:51: warning: format '%ld' expects argument of type 'long int', but argument 4 has type 'int64_t {aka long long int}' [-Wformat=]
      copied = snprintf(str_value, MAX_STR_CONST, "%ld", value);
                                                  ^
  ```

+++ src/rocto/copy_bind_parameter.c
@@ -61,20 +61,20 @@ char *copy_binary_parameter(RoctoSession *session, Bind *bind, const int cur_par
 	switch (bind->parms[cur_parm].length) {
 		case 1:
 			value = bin_to_char(bind->parms[cur_parm].value);
-			copied = snprintf(str_value, MAX_STR_CONST, "%ld", value);
+			copied = snprintf(str_value, MAX_STR_CONST, "%lld", (long long int)value);
 			break;
 		case 2:
 			value = bin_to_int16(bind->parms[cur_parm].value);
-			copied = snprintf(str_value, MAX_STR_CONST, "%ld", value);
+			copied = snprintf(str_value, MAX_STR_CONST, "%lld", (long long int)value);
 			break;
 		case 4:
 			value = bin_to_int32(bind->parms[cur_parm].value);
-			copied = snprintf(str_value, MAX_STR_CONST, "%ld", value);
+			copied = snprintf(str_value, MAX_STR_CONST, "%lld", (long long int)value);
 			break;
 		case 8:
 			// This covers the OID case, as it is just an integer
 			value = bin_to_int64(bind->parms[cur_parm].value);
-			copied = snprintf(str_value, MAX_STR_CONST, "%ld", value);
+			copied = snprintf(str_value, MAX_STR_CONST, "%lld", (long long int)value);
 			break;
 		case 16:
 			bin_to_uuid(bind->parms[cur_parm].value, uuid, MAX_STR_CONST);
